chore(connect): updatae deviceAuthenticityConfig for automation
