+++ packages/connect/src/api/firmware/verifyAuthenticityProof.ts
@@ -2,9 +2,9 @@ import * as crypto from 'crypto';
 import { bufferUtils } from '@trezor/utils';
 
 import { PROTO } from '../../constants';
-import { DeviceAuthenticityConfig } from '../../data/deviceAuthenticityConfig';
 import { AuthenticateDeviceResult } from '../../types/api/authenticateDevice';
 import { parseCertificate, fixSignature } from './x509certificate';
+import { DeviceAuthenticityConfig } from '../../data/deviceAuthenticityConfigTypes';
 
 // There is incomparability in results between nodejs and window SubtleCrypto api.
 // window.crypto.subtle.importKey (CryptoKey) cannot be used by `crypto-browserify`.Verify
